Move traits from Traits folder to Contracts folder
