+++ src/Contracts/EventManagerAwareTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Canvas\Traits;
+namespace Canvas\Contracts;
 
 use Phalcon\Di;
 use Phalcon\Events\ManagerInterface as EventsManager;
