+++ src/Contracts/SubscriptionPlanLimitTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Canvas\Traits;
+namespace Canvas\Contracts;
 
 use function Baka\getShortClassName;
 use Canvas\Exception\SubscriptionPlanLimitException;
