+++ src/Contracts/TokenTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Canvas\Traits;
+namespace Canvas\Contracts;
 
 use Lcobucci\JWT\Parser;
 use Lcobucci\JWT\Token;
