+++ src/Middleware/TokenBase.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Canvas\Middleware;
 
-use Canvas\Traits\TokenTrait;
-use Canvas\Traits\SubscriptionPlanLimitTrait;
+use Canvas\Contracts\TokenTrait;
+use Canvas\Contracts\SubscriptionPlanLimitTrait;
 use Phalcon\Mvc\Micro\MiddlewareInterface;
 
 /**
