+++ src/Models/Companies.php
@@ -8,8 +8,8 @@ use Baka\Blameable\BlameableTrait;
 use Baka\Contracts\Database\HashTableTrait;
 use Baka\Contracts\EventsManager\EventManagerAwareTrait;
 use Baka\Http\Exception\InternalServerErrorException;
-use Canvas\Traits\FileSystemModelTrait;
-use Canvas\Traits\UsersAssociatedTrait;
+use Canvas\Contracts\FileSystemModelTrait;
+use Canvas\Contracts\UsersAssociatedTrait;
 use Carbon\Carbon;
 use Exception;
 use Phalcon\Di;
