+++ src/Models/Users.php
@@ -13,9 +13,9 @@ use Baka\Hashing\Keys;
 use Baka\Hashing\Password;
 use Baka\Validations\PasswordValidation;
 use Canvas\Auth\App as AppAuth;
-use Canvas\Traits\FileSystemModelTrait;
-use Canvas\Traits\PermissionsTrait;
-use Canvas\Traits\SubscriptionPlanLimitTrait;
+use Canvas\Contracts\FileSystemModelTrait;
+use Canvas\Contracts\PermissionsTrait;
+use Canvas\Contracts\SubscriptionPlanLimitTrait;
 use Carbon\Carbon;
 use Exception;
 use Phalcon\Di;
